feat: expose job management tools to agent sessions
Extract the 13 job tools (definition CRUD, spawn, run control) from
callboard-tools.ts into a shared job-management-tools.ts builder and
inject them on the "callboard" agent server alongside deploy_agent,
create_cron_job, etc. Agent sessions now skip the copy on
callboard-tools (includeJobTools flag in claude.ts) so each session
sees exactly one set; regular chats are unchanged.

Agents stamp createdBy {kind:"agent", ref:alias} on definitions and
approvals record via:"agent". mcp-tool-registry job metadata is
factored into a server-parameterized factory so the manifest lists
the tools under the right server per context.
